Change behavior of `--drop` argument for betterbib format.
When specifying the `--drop` argument as the last argument before the
actual input files, all inputfiles are understood as fields to drop.
This is particularly annoying when using betterbib as a pre-commit hook
when one does not want to add another argument like `-b` between the
fields to drop and the input files.

This change requires each field to drop to be passed individually using
`--drop`, e.g. `betterbib -i --drop url --drop file test.bib`.